Some structs moved to mod.t
